Modernize focus history coverage
