+++ Packages/macOS/CmuxSettings/Sources/CmuxSettings/Keys/AppCatalogSection.swift
@@ -70,6 +70,8 @@ public struct AppCatalogSection: SettingCatalogSection {
         userDefaultsKey: "paneFirstClickFocus.enabled"
     )
 
+    /// Whether focus history includes pane and tab changes within a workspace.
+    /// Defaults to workspace-only navigation.
     public let focusHistoryIncludesPanesAndTabs = DefaultsKey<Bool>(
         id: "app.focusHistoryIncludesPanesAndTabs",
         defaultValue: false,
