+++ cmuxTests/TabManagerSessionSnapshotTests.swift
@@ -81,76 +81,6 @@ final class TabManagerSessionSnapshotTests: XCTestCase {
         XCTAssertTrue(manager.canNavigateForward)
     }
 
-    func testFocusHistoryWorkspacesOnlySettingSkipsPanelsInCurrentWorkspace() throws {
-        let suiteName = "TabManagerSessionSnapshotTests.focusHistory.\(UUID().uuidString)"
-        let defaults = try XCTUnwrap(UserDefaults(suiteName: suiteName))
-        defer { defaults.removePersistentDomain(forName: suiteName) }
-        let settings = UserDefaultsSettingsClient(defaults: defaults)
-        settings.set(false, for: SettingCatalog().app.focusHistoryIncludesPanesAndTabs)
-        let manager = TabManager(settings: settings)
-        let firstWorkspace = try XCTUnwrap(manager.selectedWorkspace)
-        let pane = try XCTUnwrap(firstWorkspace.bonsplitController.allPaneIds.first)
-        let firstPanelId = try XCTUnwrap(firstWorkspace.focusedPanelId)
-        let secondPanelId = try XCTUnwrap(firstWorkspace.newTerminalSurface(inPane: pane, focus: true)?.id)
-
-        firstWorkspace.focusPanel(firstPanelId)
-        firstWorkspace.focusPanel(secondPanelId)
-
-        XCTAssertFalse(manager.canNavigateBack)
-        XCTAssertFalse(manager.navigateBack())
-
-        let secondWorkspace = manager.addWorkspace(select: true)
-        XCTAssertTrue(manager.navigateBack())
-        XCTAssertEqual(manager.selectedTabId, firstWorkspace.id)
-        XCTAssertTrue(manager.navigateForward())
-        XCTAssertEqual(manager.selectedTabId, secondWorkspace.id)
-    }
-
-    func testFocusHistoryScopeChangeInvalidatesAvailability() throws {
-        let suiteName = "TabManagerSessionSnapshotTests.focusHistoryScopeChange.\(UUID().uuidString)"
-        let defaults = try XCTUnwrap(UserDefaults(suiteName: suiteName))
-        defer { defaults.removePersistentDomain(forName: suiteName) }
-        let settings = UserDefaultsSettingsClient(defaults: defaults)
-        let scopeKey = SettingCatalog().app.focusHistoryIncludesPanesAndTabs
-        settings.set(true, for: scopeKey)
-        let manager = TabManager(settings: settings)
-        let workspace = try XCTUnwrap(manager.selectedWorkspace)
-        let pane = try XCTUnwrap(workspace.bonsplitController.allPaneIds.first)
-        let firstPanelId = try XCTUnwrap(workspace.focusedPanelId)
-        let secondPanelId = try XCTUnwrap(workspace.newTerminalSurface(inPane: pane, focus: true)?.id)
-        workspace.focusPanel(firstPanelId)
-        workspace.focusPanel(secondPanelId)
-        XCTAssertTrue(manager.canNavigateBack)
-
-        let enabledRevision = manager.focusHistoryRevision
-        settings.set(false, for: scopeKey)
-        NotificationCenter.default.post(name: UserDefaults.didChangeNotification, object: defaults)
-        XCTAssertGreaterThan(manager.focusHistoryRevision, enabledRevision)
-        XCTAssertFalse(manager.canNavigateBack)
-
-        let disabledRevision = manager.focusHistoryRevision
-        settings.set(true, for: scopeKey)
-        NotificationCenter.default.post(name: UserDefaults.didChangeNotification, object: defaults)
-        XCTAssertGreaterThan(manager.focusHistoryRevision, disabledRevision)
-        XCTAssertTrue(manager.canNavigateBack)
-    }
-
-    func testRestoredWorkspaceDockUsesInjectedFocusHistorySetting() throws {
-        let suiteName = "TabManagerSessionSnapshotTests.restoredFocusHistoryScope.\(UUID().uuidString)"
-        let defaults = try XCTUnwrap(UserDefaults(suiteName: suiteName))
-        defer { defaults.removePersistentDomain(forName: suiteName) }
-        let settings = UserDefaultsSettingsClient(defaults: defaults)
-        settings.set(true, for: SettingCatalog().app.focusHistoryIncludesPanesAndTabs)
-        let source = TabManager(settings: settings)
-        let snapshot = source.sessionSnapshot(includeScrollback: false)
-
-        let restored = TabManager(settings: settings)
-        restored.restoreSessionSnapshot(snapshot)
-
-        let restoredWorkspace = try XCTUnwrap(restored.tabs.first)
-        XCTAssertTrue(restoredWorkspace.dockSplit.focusHistoryIncludesPanesAndTabs)
-    }
-
     func testFocusHistoryBackFallsBackWhenRecordedPanelWasClosed() throws {
         let manager = TabManager()
         let firstWorkspace = try XCTUnwrap(manager.selectedWorkspace)
